Separate jvm and js implementation; fix test class
