+++ modules/core/src/main/scala/io/chrisdavenport/fuuid/FUUID.scala
@@ -4,7 +4,6 @@ import cats._
 import cats.implicits._
 import cats.effect.Sync
 import java.util.UUID
-import java.security.MessageDigest
 
 import scala.reflect.macros.blackbox
 
@@ -76,48 +75,11 @@ object FUUID {
   }
 
   /**
-    * Creates a new name-based UUIDv5
+    * Creates a new name-based UUIDv5.
+    * NOTE: Not implemented for Scala.js!
     **/ 
   def nameBased[F[_]](namespace: FUUID, name: String)(implicit AE: ApplicativeError[F, Throwable]): F[FUUID] =
-    Either
-      .catchNonFatal(
-        MessageDigest
-          .getInstance("SHA-1")
-          .digest(
-            uuidBytes(namespace) ++
-              name.getBytes("UTF-8")
-          )
-      )
-      .map { bs =>
-        val cs = bs.take(16) // Truncate 160 bits (20 bytes) SHA-1 to fit into our 128 bits of space
-        cs(6) = (cs(6) & 0x0f).asInstanceOf[Byte] /* clear version                */
-        cs(6) = (cs(6) | 0x50).asInstanceOf[Byte] /* set to version 5, SHA1 UUID  */
-        cs(8) = (cs(8) & 0x3f).asInstanceOf[Byte] /* clear variant                */
-        cs(8) = (cs(8) | 0x80).asInstanceOf[Byte] /* set to IETF variant          */
-        cs
-      }
-      .flatMap(
-        bs =>
-          Either.catchNonFatal {
-            val bb = java.nio.ByteBuffer.allocate(java.lang.Long.BYTES)
-            bb.put(bs, 0, 8)
-            bb.flip
-            val most = bb.getLong
-            bb.flip
-            bb.put(bs, 8, 8)
-            bb.flip
-            val least = bb.getLong
-            new FUUID(new UUID(most, least))
-          }
-      )
-      .liftTo[F]
-
-  private def uuidBytes(fuuid: FUUID): Array[Byte] = {
-    val bb = java.nio.ByteBuffer.allocate(2 * java.lang.Long.BYTES)
-    bb.putLong(fuuid.uuid.getMostSignificantBits)
-    bb.putLong(fuuid.uuid.getLeastSignificantBits)
-    bb.array
-  }
+    PlatformSpecificMethods.nameBased[F](namespace, name, AE)
 
   /**
     * A Home For functions we don't trust
