+++ modules/core/src/test/scala/io/chrisdavenport/fuuid/FUUIDSpec.scala
@@ -4,7 +4,7 @@ import java.util.UUID
 import cats.effect.IO
 import org.specs2._
 
-object FUUIDSpec extends mutable.Specification with ScalaCheck {
+class FUUIDSpec extends mutable.Specification with ScalaCheck {
 
   "FUUID.fromString" should {
     "Fail when parsing an invalid string" in {
@@ -20,28 +20,6 @@ object FUUIDSpec extends mutable.Specification with ScalaCheck {
         .map(FUUID.fromString)
         .map(_.isRight)
         .unsafeRunSync must_=== true
-    }
-  }
-
-  "FUUID.nameBased" should {
-    "produce a valid UUID" in {
-      (for {
-        namespace <- FUUID.randomFUUID[IO]
-        namebased <- FUUID.nameBased[IO](namespace, "What up yo!")
-        parsed <- FUUID.fromStringF[IO](namebased.toString)
-      } yield parsed)
-        .attempt
-        .unsafeRunSync
-        .isRight must_=== true
-    }
-
-    "conform to an example" in {
-      val namespace = FUUID.fromStringF[IO]("dc79a6bc-de3c-5bc7-a877-712bea708d8f").unsafeRunSync()
-      val name = "What up yo!"
-
-      val expected = FUUID.fromStringF[IO]("1cce4593-d217-5b33-be0d-2e81462e79d3").unsafeRunSync()
-
-      FUUID.nameBased[IO](namespace, name).unsafeRunSync must_=== expected
     }
   }
